requirements.txt
azure-identity
openai>=1.109.1
python-dotenv
pydantic
rich
azure-ai-inference
dotenv-azd
aiohttp
faker
autogen-agentchat
autogen-ext[openai]
azure-ai-inference==1.0.0b9
openai-agents
semantic-kernel
langgraph
langchain==1.0.0a5
langchain_openai
pydantic-ai==1.0.8
llama-index
llama-index-llms-azure-openai
llama-index-llms-openai-like
llama-index-embeddings-azure-openai
smolagents
langchain-mcp-adapters
agent-framework-core
